% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biGram_pt.R
\name{biGram_pt}
\alias{biGram_pt}
\title{Bigram probability for Portuguese}
\usage{
biGram_pt(word = "")
}
\arguments{
\item{word}{A possible string in Portuguese in its phonemic form without syllabification or stress. The only diacritic that should be used is the tilde for nasals, e.g., ã.}
}
\value{
The phonemic transcription for the string in question
}
\description{
Given a phonemically transcribed string, the function returns its bigram probability in log using the lexicon in the Portuguese Stress Lexicon as reference
}
\examples{
ipa_pt(word = "palado");
ipa_pt(word = "sɛtiko");
ipa_pt(word = "mãw̃");
}
