% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consonants_en}
\alias{consonants_en}
\title{Data: consonant inventory for English}
\format{
A vector containing the consonants and glides in English
}
\usage{
data(consonants_en)
}
\description{
The inventory of consonants and glides in English
}
\seealso{
\code{\link{getFeat}}, \code{\link{getPhon}}
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
