% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consonants_fr}
\alias{consonants_fr}
\title{Consonant inventory for French}
\format{
A vector containing the consonants and glides in French
}
\usage{
data(consonants_fr)
}
\description{
The inventory of consonants and glides in French
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
