% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consonants_it}
\alias{consonants_it}
\title{Data: consonant inventory for Italian}
\format{
A vector containing the consonants and glides in Italian
}
\usage{
data(consonants_it)
}
\description{
The inventory of consonants and glides in Italian
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
