% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consonants_pt}
\alias{consonants_pt}
\title{Consonant inventory for Portuguese}
\format{
A vector containing the consonants and glides in Portuguese
}
\usage{
consonants_pt
}
\description{
The inventory of consonants and glides in Portuguese
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
