% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consonants_sp}
\alias{consonants_sp}
\title{Data: consonant inventory for Spanish}
\format{
A vector containing the consonants and glides in Spanish
}
\usage{
data(consonants_sp)
}
\description{
The inventory of consonants and glides in Spanish
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
