% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllable_shape.R
\name{cv}
\alias{cv}
\title{Syllable shape for transcribed strings}
\usage{
cv(word = c(""), stress = FALSE)
}
\arguments{
\item{word}{A possible string in its phonemic form, i.e., the output of \code{ipa()}.}

\item{stress}{Whether the stress diacritic from \code{ipa()} should be kept in the output (default is \code{FALSE})}
}
\value{
The syllable shape of the string
}
\description{
Given a phonemically transcribed string, the function returns its syllable shape using C, V and G to represent consonants, vowels and glides, respectively.
}
\examples{
cv(word = "pa.ˈklo.do")
}
