% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dactyl_pt.R
\name{dact_pt}
\alias{dact_pt}
\title{Dactylic lowering helper function}
\usage{
dact_pt(word = "")
}
\arguments{
\item{word}{A possible string in Portuguese in its phonemic form}
}
\value{
The transcription with dactylic lowering if applicable
}
\description{
Given a string, the function tests if the string is LLL and has antepenult stress,
in which case the antepenultimate vowel height is checked and lowered if needed.
}
\examples{
dact_pt(word = "ˈpe.te.le");
}
