% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demi.R
\name{demi}
\alias{demi}
\title{Demi syllable extractor}
\usage{
demi(word = c(), d = 1)
}
\arguments{
\item{word}{A syllabified and phonemically transcribed word; vectors are accepted}

\item{d}{Whether the function should return first \code{(1)} or second \code{(2)} demisyllable}
}
\value{
The demisyllable of interest
}
\description{
Given a string, the function returns its IPA transcription with stress and syllabification.
Stress is assigned on the basis of the Portuguese Stress Lexicon for existing words, or lexical regularities
and probabilistic distributions in certain cases for hypothetical words.
}
\examples{
demi(word = c("kram.pjo", "tlons.tri"), d = 1)
}
