% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dePluralize_pt.R
\name{dePlu_pt}
\alias{dePlu_pt}
\title{Function to depluralize existing words}
\usage{
dePlu_pt(word = "")
}
\arguments{
\item{word}{A word in its orthographic form}
}
\value{
The singular form of the word
}
\description{
Extracts the singular form of a plural word
}
\examples{
dePlu_pt(word = "variedades")
}
