% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diacritic_pt.R
\name{diacritic_pt}
\alias{diacritic_pt}
\title{Diacritic detector in Portuguese orthography}
\usage{
diacritic_pt(word)
}
\arguments{
\item{word}{A possible string in Portuguese in its orthographic form}
}
\value{
The position of a stress diacritic (if present) counting from the
right edge of the word. The function returns NA otherwise. If multiple stress diacritics are provided,
or if a diacritic is outside the stress window, an NA is returned.
}
\description{
This is a helper function that detects whether a stress diacritic is present in the input,
which allows for stress assignment to be more accurate in so-called exceptional stress words
in the language (e.g., all proparoxytones are orthographically accented). A stress diacritic
is an acute accent, a grave accent, or a circumflex accent. The nasal diacritic are not included
as it is already accounted for by ipa_pt(). Grave accents are included to accommodate potential
typos or OCR errors for digitized inputs.
}
\examples{
ipa_pt(word = "pálado");
}
