% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{en_lex}
\alias{en_lex}
\title{Data: English CMU-derived lexicon}
\format{
A data frame with 133,854 rows and 6 variables:
\describe{
  \item{\code{word}}{English orthographic form}
  \item{\code{alt}}{CMU pronunciation index, where 0 is the default entry}
  \item{\code{arpabet}}{ARPAbet transcription from CMUdict}
  \item{\code{ipa_plain}}{IPA transcription without syllable separators}
  \item{\code{ipa_syll}}{IPA transcription with syllabification and stress}
  \item{\code{stress_pattern}}{Stress digits extracted from vowel tokens}
}
}
\usage{
data(en_lex)
}
\description{
A precomputed English lexicon derived from CMUdict and used by
\code{ipa(lg = "English")}. It contains the orthographic form, the default
ARPAbet entry used by the package, and precomputed IPA transcriptions with
stress and syllabification.
}
\seealso{
\code{\link{add_lex_en}}
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
