% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{features_Hayes_2009}
\alias{features_Hayes_2009}
\title{Data: contrastive features from Hayes (2009)
The complete table of features found in Hayes's 2009
Introductory Phonology}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 132 rows and 28 columns.
}
\usage{
data(features_Hayes_2009)
}
\description{
Data: contrastive features from Hayes (2009)
The complete table of features found in Hayes's 2009
Introductory Phonology
}
\author{
Guilherme D. Garcia
}
\keyword{datasets}
