% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pt.R
\name{gen_pt}
\alias{gen_pt}
\title{Word generator for Portuguese}
\usage{
gen_pt(profile = "LLL")
}
\arguments{
\item{profile}{The weight profile of the desired string using Ls or Hs}
}
\value{
The IPA transcription of said string
}
\description{
Returns IPA phonemic transcription for a nonce word given a specific weight profile
}
\examples{
gen_pt(profile = "HLL")
}
