% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSyl.R
\name{getSyl}
\alias{getSyl}
\title{Function to extract syllables from words}
\usage{
getSyl(word = c("pa.la.do"), pos = 1, from_right = TRUE, syl = "\\\\.")
}
\arguments{
\item{word}{The strings of interest must be syllabified}

\item{pos}{The target syllable counting from the right edge of the word}

\item{from_right}{Whether to parse from the right (default is \code{TRUE}).}

\item{syl}{The symbol used for syllable boundaries (a period is used as the default)}
}
\value{
The desired syllable if it exists. The function returns \code{NA} otherwise
}
\description{
Extracts a given syllable from syllabified strings in a vector.
}
\examples{
getSyl(word = c("kom.pu.ta.dor", "pin.to.de"), pos = 2)
}
