% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pt_vec.R
\name{getWeight_pt}
\alias{getWeight_pt}
\title{Syllable weight labeller}
\usage{
getWeight_pt(word = c("kom.pu.ta.ˈdoɾ"))
}
\arguments{
\item{word}{The vector with strings of interest using IPA phonemic transcription, already syllabified and stressed}
}
\value{
The sequence of Ls and Hs based on the word's weight profile given the phonology of Portuguese
}
\description{
Labels a given string in Portuguese according to its weight profile using Ls and Hs.
}
\examples{
getWeight_pt(word = c("kom.pu.ta.ˈdoɾ", "ka.ˈloɾ.pe"));
}
