% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pt_vec.R
\name{getWeight_pt_vec}
\alias{getWeight_pt_vec}
\title{Vectorized syllable weight labeller}
\usage{
getWeight_pt_vec(word = c("kom.pu.ta.ˈdoɾ"))
}
\arguments{
\item{word}{The vector with strings of interest using IPA phonemic transcription, already syllabified and stressed}
}
\value{
The sequence of Ls and Hs based on the word's weight profile given the phonology of Portuguese
}
\description{
Labels a given string in Portuguese according to its weight profile using Ls and Hs.
Note that, although simple plural -s is taken into account, the function assumes the string provided is monomorphemic.
For a simplified but vectorized version of the function, use getWeight_pt_simple().
}
\examples{
getWeight_pt_vec(word = c("kom.pu.ta.ˈdoɾ", "ka.ˈloɾ.pe"));
}
