% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_sp_vec.R
\name{getWeight_sp}
\alias{getWeight_sp}
\title{Syllable weight labeller for Spanish}
\usage{
getWeight_sp(word = c("kom.pu.ta.ˈdoɾ"))
}
\arguments{
\item{word}{The vector with strings of interest using IPA phonemic transcription, already syllabified and stressed}
}
\value{
The sequence of Ls and Hs based on the word's weight profile given the phonology of Spanish
}
\description{
Labels a given string in Portuguese according to its weight profile using Ls and Hs.
}
\examples{
getWeight_sp(word = c("kom.pu.ta.ˈdoɾ", "ka.ˈloɾ.pe"));
}
