% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa.R
\name{ipa}
\alias{ipa}
\title{IPA transcriber}
\usage{
ipa(word, lg = "Portuguese", narrow = FALSE)
}
\arguments{
\item{word}{A possible string in its orthographic form}

\item{lg}{Language. Currently, only Portuguese and Spanish are supported}

\item{narrow}{Boolean. Whether a narrow transcription is desired (default is \code{FALSE})
Narrow transcription is only available for Portuguese inputs and will be ignored for other languages}
}
\value{
The phonemic transcription for the string in question
}
\description{
Given a string, the function returns its IPA transcription with stress and syllabification
}
\examples{
ipa(word = "palado");
}
