% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa2typst.R
\name{ipa2typst}
\alias{ipa2typst}
\title{Phonokit IPA translator}
\usage{
ipa2typst(string, pre = "#ipa(\\"", post = "\\")")
}
\arguments{
\item{string}{The phonemically transcribed sequence from a function such as \code{ipa()}}

\item{pre}{Prefix for transcription. Defaults to \code{'#ipa("'}}

\item{post}{Suffix for transcription. Defaults to \code{'")'}}
}
\value{
A string with phonokit notation (printed to the console and returned invisibly)
}
\description{
Translates a phonemically transcribed sequence (IPA Unicode) into
\href{https://github.com/guilhermegarcia/phonokit}{phonokit} notation
for use in Typst documents.
Any token that requires a backslash command is automatically
surrounded by spaces, as required by phonokit's parser.
}
\examples{
"bonito" |> ipa(lg = "Portuguese") |> ipa2typst()
"chuva" |> ipa(lg = "Portuguese") |> ipa2typst()
}
