% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_words_en.R
\name{ipa_en_test}
\alias{ipa_en_test}
\title{IPA tester for \code{ipa()} function in Fonology package}
\usage{
ipa_en_test()
}
\value{
A list of words broadly transcribed using \code{ipa(lg = "English")}
}
\description{
Returns a list of English words. No arguments needed.
}
\examples{
ipa_en_test()
}
