% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_words_fr.R
\name{ipa_fr_test}
\alias{ipa_fr_test}
\title{IPA tester for \code{ipa()} function in Fonology package}
\usage{
ipa_fr_test()
}
\value{
A list of words broadly and narrowly transcribed using \code{ipa(lg = "French")}
}
\description{
Returns a list of words. No arguments needed.
}
\examples{
ipa_fr_test()
}
