% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_words_it.R
\name{ipa_it_test}
\alias{ipa_it_test}
\title{IPA tester for \code{ipa()} function in Fonology package (Italian)}
\usage{
ipa_it_test()
}
\value{
A list of words transcribed using \code{ipa(lg = "Italian")}
}
\description{
Returns a list of Italian words transcribed using \code{ipa(lg = "Italian")}.
No arguments needed. Words are chosen to exercise the main phonological
patterns of Standard Italian: palatals, geminates, final stress, glides,
and intervocalic voicing.
}
\examples{
ipa_it_test()
}
