% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPA_sp.R
\name{ipa_sp}
\alias{ipa_sp}
\title{IPA transcriber for Spanish}
\usage{
ipa_sp(word = "comportamento")
}
\arguments{
\item{word}{A possible string in Portuguese in its orthographic form}
}
\value{
The phonemic transcription for the string in question
}
\description{
Given a string, the function returns its IPA transcription with stress and syllabification.
}
\examples{
ipa_sp(word = "comportamento");
}
