% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_words_sp.R
\name{ipa_sp_test}
\alias{ipa_sp_test}
\title{IPA tester for \code{ipa()} function in Fonology package}
\usage{
ipa_sp_test()
}
\value{
A list of words broadly and narrowly transcribed using \code{ipa(lg = "Spanish")}
}
\description{
Returns a list of words. No arguments needed.
}
\examples{
ipa_sp_test();
}
