% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{it_lex_fix}
\alias{it_lex_fix}
\title{Data: Italian shipped lexicon corrections}
\format{
A named character vector
}
\usage{
data(it_lex_fix)
}
\description{
A named character vector of maintainer-curated stress and mid-vowel quality
corrections distributed with the package and used by
\code{ipa(lg = "Italian")}. Keys are plain orthographic forms; values are
diacritized forms. Merged at run time with \code{\link{it_lex_user}}, which
takes priority.
}
\seealso{
\code{\link{promote_lex}}, \code{\link{add_lex_it}}
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
