% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lex_en_cmu_complement}
\alias{lex_en_cmu_complement}
\title{Data: English CMU-complement lexicon}
\format{
A data frame with 2 variables:
\describe{
  \item{\code{word}}{English orthographic form}
  \item{\code{ipa}}{IPA transcription with syllabification and stress}
}
}
\usage{
data(lex_en_cmu_complement)
}
\description{
A curated English lexicon used by \code{ipa(lg = "English")} after
user overrides and before the CMU-derived lexicon. It stores complete IPA
transcriptions with stress and syllabification for forms that are absent
from CMUdict or need package-level supplementation.
}
\seealso{
\code{\link{add_lex_en}}, \code{\link{ipa}}
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
