% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_conversion.R
\name{meanAge}
\alias{meanAge}
\title{Mean age calculator}
\usage{
meanAge(age, sep = ";")
}
\arguments{
\item{age}{Age in yy;mm format}

\item{sep}{Separator for age. Default is ";"}
}
\value{
Mean age in yy;mm format using the separator specified in `sep`
}
\description{
Calculates mean age based on yy;mm format, commonly used in language acquisition studies. `NA`s are ignored
}
\examples{
meanAge(age = c("02;06", "03;04", NA))
meanAge(age = c("05:03", "04:07"), sep = ":")
}
