% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSonDisp.R
\name{meanSonDisp}
\alias{meanSonDisp}
\title{Mean sonority dispersion calculator}
\usage{
meanSonDisp(word = c(), phonemic = F)
}
\arguments{
\item{word}{A word in its orthographic (only for Portuguese) or phonemic form}

\item{phonemic}{Whether the input is phonemically transcribed (words in languages other than Portuguese must be set to TRUE)
and syllabified. Default is FALSE}
}
\value{
The mean sonority dispersion score
}
\description{
Given a word or a vector of words, the function returns
the mean sonority dispersion score based on the
sonority sequencing principle
}
\examples{
meanSonDisp(word = c("combrade", "prondo", "clauspricos"));
}
