% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrow_pt.R
\name{narrow_pt}
\alias{narrow_pt}
\title{Narrow IPA transcriber for Portuguese}
\usage{
narrow_pt(word = "")
}
\arguments{
\item{word}{The string of interest is an output of ipa_pt()}
}
\value{
The IPA transcription of said string with surface-level adjustments
}
\description{
Returns IPA phonemic transcription for a given string that is already phonemically transcribed
}
\examples{
narrow_pt(word = "ˈpsi.ko");
}
