% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhg.R
\name{nhg}
\alias{nhg}
\title{Noisy Harmonic Grammar simulation}
\usage{
nhg(
  tableau,
  weights,
  input_col = "input",
  output_col = "output",
  n_sim = 100,
  noise_sd = 1,
  seed = NULL
)
}
\arguments{
\item{tableau}{The data object (data frame or tibble) containing an \code{input} column
and an \code{output} column, in addition to all relevant constraints, their respective
violations}

\item{weights}{A vector with the weights for the constraints in \code{tableau}}

\item{input_col}{The name of the input column in \code{tableau} (default: \code{input})}

\item{output_col}{The name of the output column in \code{tableau} (default: \code{output})}

\item{n_sim}{The number of simulations to generate the probabilities for all candidates}

\item{noise_sd}{The standard deviation of the Gaussian distribution from which noise will be added to constraint weights}

\item{seed}{For reproducibility}
}
\description{
Given an object with one or more tableaux and an object with the weights for all the constraints in question, the function returns the probabilities for each candidate by adding random noise from a Gaussian distribution to each constraint weight.
}
\examples{
tableaux <- tibble::tibble(
  input = rep(c("/input_A/", "/input_B/"), each = 3),
  output = rep(c("[output_1]", "[output_2]", "[output_3]"), times = 2),
  C1 = c(-1, 0, 0, -2, 0, -1),
  C2 = c(0, -2, -1, 0, -1, -1),
  C3 = c(0, 0, -1, 0, 0, -1)
)

weights <- c(15, 8, 8)

nhg(tableau = tableaux, weights, n_sim = 100, noise_sd = 1)
}
