% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigram_viz.R
\name{plot_biGrams}
\alias{plot_biGrams}
\title{Visualizing bigrams}
\usage{
plot_biGrams(bigrams, type = "lollipop")
}
\arguments{
\item{bigrams}{A tibble in the exact format used in biGram_tbl()}

\item{type}{The type of plot desired: heatmap or lollipop (default)}
}
\value{
A ggplot2 plot with the most common bigrams
}
\description{
Given a tibble with phonotactic bigrams (result of biGram_tbl()), the function plots the n most common bigrams
}
\examples{
plot_biGrams(bigrams = "")
}
