% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigram_viz.R
\name{plotnGrams}
\alias{plotnGrams}
\title{Visualizing nGrams}
\usage{
plotnGrams(ngrams, type = "lollipop", n = 5)
}
\arguments{
\item{ngrams}{A tibble in the exact format used in biGram_tbl()}

\item{type}{The type of plot desired: heatmap or lollipop (default)}

\item{n}{The number of ngrams to plot}
}
\value{
A ggplot2 plot with the most common ngrams
}
\description{
Given a tibble with phonotactic ngrams (result of nGram_tbl(..., n = 2)), the function plots the n most common ngrams
}
\examples{
plotnGrams(ngrams = "")
}
