% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhg.R
\name{plotNhg}
\alias{plotNhg}
\title{Plotting probabilities as a function of sigma in harmonic grammar simulations}
\usage{
plotNhg(tableau, weights)
}
\arguments{
\item{tableau}{The data object (data frame or tibble) containing an \code{input} column
and an \code{output} column, in addition to all relevant constraints, their respective
violations}

\item{weights}{A vector with the weights for the constraints in \code{tableau}}
}
\description{
Given an object with one or more tableaux and an object with the weights for all the constraints in question, the function creates a figure using \code{ggplot2} showing how candidate probabilities change as a function of the standard deviation of the Gaussian distribution from which the noise is derived. Sigmas 1:5 are used to illustrate the effect of different standard deviation values.
}
\examples{
tableaux <- tibble::tibble(
  input = rep(c("/input_A/", "/input_B/"), each = 3),
  output = rep(c("[output_1]", "[output_2]", "[output_3]"), times = 2),
  C1 = c(-1, 0, 0, -2, 0, -1),
  C2 = c(0, -2, -1, 0, -1, -1),
  C3 = c(0, 0, -1, 0, 0, -1)
)

weights <- c(15, 8, 8)

plotNhg(tableau = tableaux, weights)
}
