% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_lex_en.R
\name{remove_lex_en}
\alias{remove_lex_en}
\title{Remove words from the English user lexicon}
\usage{
remove_lex_en(words)
}
\arguments{
\item{words}{A character vector of words to remove.}
}
\value{
Invisibly returns the updated IPA lexicon.
}
\description{
Removes one or more words from the English IPA-override lexicon.
}
\seealso{
\code{\link{add_lex_en}}, \code{\link{export_lex}}
}
