% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_lex_fr.R
\name{remove_lex_fr}
\alias{remove_lex_fr}
\title{Remove words from the French user lexicon}
\usage{
remove_lex_fr(words)
}
\arguments{
\item{words}{A character vector of words to remove (plain orthographic
forms).}
}
\value{
Invisibly returns the updated IPA-override lexicon.
}
\description{
Removes one or more words from the French IPA-override lexicon.
}
\seealso{
\code{\link{add_lex_fr}}, \code{\link{export_lex}}
}
