% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp.R
\name{sonDisp}
\alias{sonDisp}
\title{Sonority dispersion calculator}
\usage{
sonDisp(demi = "")
}
\arguments{
\item{demi}{A demi syllable; see function \code{demi()}}
}
\value{
The sonority dispersion score
}
\description{
Given a demisyllable, the function returns its sonority dispersion score based on the
sonority sequencing principle.
}
\examples{
sonDisp(demi = "tra")
}
