% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_ipa_fix}
\alias{sp_ipa_fix}
\title{Data: Spanish shipped IPA overrides}
\format{
A named character vector
}
\usage{
data(sp_ipa_fix)
}
\description{
A named character vector of maintainer-curated IPA overrides distributed
with the package. Merged at run time with \code{\link{sp_ipa_lex}}, which
takes priority.
}
\seealso{
\code{\link{promote_lex}}, \code{\link{add_lex_sp}}
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
