% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_lex_fix}
\alias{sp_lex_fix}
\title{Data: Spanish shipped lexicon corrections}
\format{
A named character vector
}
\usage{
data(sp_lex_fix)
}
\description{
A named character vector of maintainer-curated corrections distributed with
the package and used by \code{ipa(lg = "Spanish")}. Keys are plain
orthographic forms; values are diacritized forms. Merged at run time with
\code{\link{sp_lex_user}}, which takes priority.
}
\seealso{
\code{\link{promote_lex}}, \code{\link{add_lex_sp}}
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
