% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spondaic_pt.R
\name{spond_pt}
\alias{spond_pt}
\title{Spondaic lowering helper function}
\usage{
spond_pt(word = "")
}
\arguments{
\item{word}{A possible string in Portuguese in its phonemic form}
}
\value{
The transcription with spondaic lowering if applicable
}
\description{
Given a string, the function tests if the string is LH and has penult stress,
in which case the penultimate vowel height is checked and lowered if needed.
}
\examples{
ipa_pt(word = "ˈpe.tel");
}
