% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_en}
\alias{stopwords_en}
\title{Data: stopwords in English}
\format{
A character vector with stopwords in English
}
\source{
<https://cran.r-project.org/web/packages/stopwords/index.html>
}
\usage{
data(stopwords_en)
}
\description{
Set of stopwords in English from the stopwords package (stopwords-iso source).
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
