% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_it}
\alias{stopwords_it}
\title{Data: stopwords in Italian}
\format{
A character vector with stopwords in Italian
}
\source{
<https://cran.r-project.org/web/packages/stopwords/index.html>
}
\usage{
data(stopwords_it)
}
\description{
Set of stopwords in Italian from the stopwords package (stopwords-iso source).
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
