% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_pt}
\alias{stopwords_pt}
\title{Data: stopwords in Portuguese}
\format{
A character vector with stopwords in Portuguese
}
\source{
<https://cran.r-project.org/web/packages/stopwords/index.html>
}
\usage{
data(stopwords_pt)
}
\description{
Set of stopwords in Portuguese from the stopwords package (stopwords-iso source).
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
