% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_sp}
\alias{stopwords_sp}
\title{Data: stopwords in Spanish}
\format{
A character vector with stopwords in Spanish
}
\source{
<https://cran.r-project.org/web/packages/stopwords/index.html>
}
\usage{
data(stopwords_sp)
}
\description{
Set of stopwords in Spanish from the stopwords package (stopwords-iso source).
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
