% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_pt.R
\name{stress_pt}
\alias{stress_pt}
\title{Stress assigner for Portuguese}
\usage{
stress_pt(word = "")
}
\arguments{
\item{word}{The string of interest using IPA phonemic transcription and already syllabified}
}
\value{
The stressed version of the string in question
}
\description{
Assigns stress to a given string
}
\examples{
stress_pt(word = "kom.pu.ta.doɾ");
}
