% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_pt_vec.R
\name{stress_pt_simple}
\alias{stress_pt_simple}
\title{Vectorized stress assigner for Portuguese}
\usage{
stress_pt_simple(word = c("ka.va.lo"))
}
\arguments{
\item{word}{A vector with possible strings in Portuguese, which must be phonemically transcribed and syllabified}
}
\value{
The stressed strings
}
\description{
Given a string, the function assigns regular stress to a vector of strings
Stress is categorically defined, so the function is more simplistic than stress_pt().
and probabilistic distributions in certain cases for hypothetical words
}
\examples{
stress_pt_simple(word = c("palado", "antedom"));
}
