% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_pt_vec.R
\name{stress_pt_vec}
\alias{stress_pt_vec}
\title{Vectorized stress assigner for Portuguese}
\usage{
stress_pt_vec(word = c("ka.va.lo"))
}
\arguments{
\item{word}{A vector with possible strings in Portuguese, which must be phonemically transcribed and syllabified}
}
\value{
The stressed strings
}
\description{
The function assigns regular to a vector of strings
Stress is categorically defined, so the function is more simplistic than stress_pt().
}
\examples{
stress_pt_vec(word = c("pa.la.do", "an.te.dom"));
}
