% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clitics_pt.R
\name{strip_clitic_pt}
\alias{strip_clitic_pt}
\title{Clitic removal for Portuguese}
\usage{
strip_clitic_pt(word = "")
}
\arguments{
\item{word}{A possible string in Portuguese in its orthographic form}
}
\value{
A word without its hyphenated clitics
}
\description{
Given a string, the function removes its hyphenated clitics.
}
\examples{
strip_clitic_pt(word = "dá-lhe");
}
