% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabify_pt.R
\name{syllabify_pt}
\alias{syllabify_pt}
\title{Syllabifier for Portuguese}
\usage{
syllabify_pt(word = "")
}
\arguments{
\item{word}{The string of interest using IPA phonemic transcription}
}
\value{
The syllabification for the string in question
}
\description{
Returns syllabification for a given string
}
\examples{
syllabify_pt(word = "komputadoɾ");
}
