% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabify_pt_vec.R
\name{syllabify_pt_vec}
\alias{syllabify_pt_vec}
\title{Vectorized syllabifier for Portuguese}
\usage{
syllabify_pt_vec(word = "")
}
\arguments{
\item{word}{The strings of interest using IPA phonemic transcription}
}
\value{
The syllabification for the strings in question
}
\description{
Returns syllabification for a given string
}
\examples{
syllabify_pt_vec(word = c("pa.la.do", "an.te.dom"));
}
