% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllCount.R
\name{syllCount}
\alias{syllCount}
\title{Syllable counter}
\usage{
syllCount(word = "", sep = "\\\\.")
}
\arguments{
\item{word}{A possible string that is already syllabified}

\item{sep}{The syllable boundary employed. By default, the
function assumes a period separates syllables, but you may want to
change that using the `sep` argument. Bear in mind that you must
follow the typical syntax used in RegEx in R for special symbols}
}
\value{
The number of syllables in the string of interest
}
\description{
The function counts the number of syllables in a given string
}
\examples{
syllCount("kan.ˈto.zo")
}
