% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tableau_maxent_1}
\alias{tableau_maxent_1}
\title{Data: sample MaxEnt tableau (1)}
\format{
A tibble to be used with the \code{maxent()} function
}
\usage{
data(tableau_nhg_2)
}
\description{
Dummy example to be used in intro to MaxEnt Grammars
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
