% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tableau_nhg_1}
\alias{tableau_nhg_1}
\title{Data: sample tableau for NHG (1)}
\format{
A tibble to be used with the \code{nhg()} function
}
\usage{
data(tableau_nhg_1)
}
\description{
Dummy example to be used in intro to Noisy Harmonic Grammars
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
