% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tableau_nhg_2}
\alias{tableau_nhg_2}
\title{Data: sample tableau for NHG (2)}
\format{
A vector containing the consonants and glides in Spanish
}
\usage{
data(tableau_nhg_2)
}
\description{
The inventory of consonants and glides in Spanish
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
