% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_words_pt.R
\name{test_pt}
\alias{test_pt}
\title{IPA tester for ipa_pt function in Fonology package}
\usage{
test_pt()
}
\value{
A list of words broadly and narrowly transcribed using ipa_pt
}
\description{
Returns a list of words. No arguments needed.
}
\examples{
;
}
