% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe_pt_vec.R
\name{transcribe_pt_vec}
\alias{transcribe_pt_vec}
\title{Vectorized IPA transcriber for Portuguese}
\usage{
transcribe_pt_vec(word = "")
}
\arguments{
\item{word}{The strings of interest}
}
\value{
The IPA transcription of said string without syllabification or stress
}
\description{
Returns IPA phonemic transcription for a given group of strings
}
\examples{
transcribe_pt_vec(word = c("pa.la.do", "an.te.dom"));
}
