% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vowels_en}
\alias{vowels_en}
\title{Data: vowel inventory for English}
\format{
A vector containing the vowels in English
}
\usage{
data(vowels_en)
}
\description{
The inventory of vowels in English, including the r-coloured vowels
\enc{ɚ}{(U+025A)} and \enc{ɝ}{(U+025D)}
}
\seealso{
\code{\link{getFeat}}, \code{\link{getPhon}}
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
