% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vowels_fr}
\alias{vowels_fr}
\title{Data: vowel inventory for French}
\format{
A vector containing the consonants and glides in French
}
\usage{
data(vowels_fr)
}
\description{
The inventory of vowels in French
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
