% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vowels_it}
\alias{vowels_it}
\title{Vowel inventory for Italian}
\format{
A vector containing the consonants and glides in Italian
}
\usage{
data(vowels_it)
}
\description{
The inventory of vowels in Italian
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
