% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vowels_pt}
\alias{vowels_pt}
\title{Vowel inventory for Portuguese}
\format{
A vector containing the consonants and glides in Portuguese
}
\usage{
vowels_pt
}
\description{
The inventory of vowels in Portuguese
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
