% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vowels_sp}
\alias{vowels_sp}
\title{Data: vowel inventory for Spanish}
\format{
A vector containing the consonants and glides in Spanish
}
\usage{
data(vowels_sp)
}
\description{
The inventory of vowels in Spanish
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
