% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pt.R
\name{weight_pt}
\alias{weight_pt}
\title{Syllable weight labeller}
\usage{
weight_pt(word = "")
}
\arguments{
\item{word}{The string of interest using IPA phonemic transcription, already syllabified and stressed}
}
\value{
The sequence of Ls and Hs based on the word's weight profile given the phonology of Portuguese
}
\description{
Labels a given string according to its weight profile using Ls and Hs
}
\examples{
weight_pt(word = "kom.pu.ta.'doɾ");
}
