% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_stress_pt.R
\name{which_stress_pt}
\alias{which_stress_pt}
\title{Stress labeller for Portuguese}
\usage{
which_stress_pt(word = "")
}
\arguments{
\item{word}{The string of interest using IPA phonemic transcription, already syllabified and stressed}
}
\value{
The sequence of Ls and Hs based on the word's weight profile given the phonology of Portuguese
}
\description{
Labels a given stressed string as follows: Final, Penult, Antepenult
}
\examples{
weight_pt(word = "kom.pu.ta.ˈdoɾ");
}
