% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pt.R
\name{wug_pt}
\alias{wug_pt}
\title{Word generator for Portuguese}
\usage{
wug_pt(profile = "LLL", palatalization = F)
}
\arguments{
\item{profile}{The weight profile of the desired string using Ls or Hs}

\item{palatalization}{Whether t and d should palatalize before i (default is FALSE)}
}
\value{
The IPA transcription of said string
}
\description{
Returns IPA phonemic transcription for a nonce word given a specific weight profile
}
\examples{
gen_pt(profile = "HLL")
}
