% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allFeatures}
\alias{allFeatures}
\title{Distinctive features}
\format{
## `allFeatures`
A data frame with with one IPA symbol per row and one feature per column
}
\source{
<https://github.com/dmort27/panphon/blob/master/panphon/data/ipa_all.csv>
}
\usage{
allFeatures
}
\description{
Comprehensive dataset with phonemes and their distinctive features.
The list is based on PanPhon, by Mortensen et al (2016), whose GitHub can be found
at https://github.com/dmort27/panphon
}
\keyword{datasets}
