% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allFeatures}
\alias{allFeatures}
\title{Data: comprehensive table with distinctive features}
\format{
A data frame with with one IPA symbol per row and one feature per column
The (abbreviated) available features are: \code{syl}, \code{son},
\code{cons}, \code{cont}, \code{DR}, \code{lat}, \code{nas}, \code{strid},
\code{vce}, \code{sg}, \code{cg}, \code{ant}, \code{cor}, \code{distr},
\code{lab}, \code{hi}, \code{lo}, \code{back}, \code{round}, \code{vel},
\code{tense}, \code{long}, \code{hitone}, \code{hireg}, \code{approx}
}
\source{
Adapted from <https://github.com/dmort27/panphon/blob/master/panphon/data/ipa_all.csv>
}
\usage{
data(allFeatures)
}
\description{
Comprehensive dataset with phonemes and their distinctive features.
The list is based on PanPhon, by Mortensen et al (2016), whose GitHub can be found
at https://github.com/dmort27/panphon
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
