% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigrams_pt}
\alias{bigrams_pt}
\title{Bigrams in Portuguese}
\format{
A tibble with bigrams, frequency, and proportions
}
\source{
<https://gdgarcia.ca/psl>
}
\usage{
bigrams_pt
}
\description{
Set of bigrams extracted from the Portuguese Stress Lexicon using the ngram package.
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
