% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanText.R
\name{cleanText}
\alias{cleanText}
\title{Text cleaner}
\usage{
cleanText(text = "")
}
\arguments{
\item{text}{A possible string or text}
}
\value{
A vector with all words in the input, stripped of punctuation
}
\description{
Given a string, the function removes punctuation, empty tokens, numbers, and normalizes lower case.
}
\examples{
cleanText(text = "Este é um texto em português? This is a text in English!");
}
