% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonology-package.R
\docType{package}
\name{Fonology-package}
\alias{Fonology}
\alias{Fonology-package}
\title{Fonology: Phonological analysis in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package contains a range of functions to help with phonological analysis and teaching.
}
\keyword{internal}
