% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonology-package.R
\docType{package}
\name{Fonology-package}
\alias{Fonology}
\alias{Fonology-package}
\title{Fonology: Phonological Analysis in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of tools to help with phonological analysis. Functions include grapheme to phoneme conversion, sonority dispersion calculation, minimal matrices for natural classes given a set of features, maxent models, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gdgarcia.ca/fonology}
}

}
\author{
\strong{Maintainer}: Guilherme D. Garcia \email{guilherme.garcia@lli.ulaval.ca} (\href{https://orcid.org/0000-0003-1412-3856}{ORCID}) [copyright holder]

}
\keyword{internal}
