% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formant_wrapper.R
\name{formants}
\alias{formants}
\title{Vowel formant wrapper for ggplot2}
\usage{
formants()
}
\value{
Axis reversal for both x and y
}
\description{
Reverses axes to accommodate F1 and F2 formants
}
\examples{
formants();
}
