% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonemes_to_features.R
\name{getFeat}
\alias{getFeat}
\title{Distinctive feature generator}
\usage{
getFeat(ph = c(), lg = "Portuguese")
}
\arguments{
\item{ph}{The phonemes of interest}

\item{lg}{The language of interest: English, French, Italian, Portuguese, Spanish}
}
\value{
The minimal matrix of features given ph and lg
}
\description{
Generates a feature matrix for a given set of phonemes in a given language
}
\examples{
getFeat(ph = c("i", "u"), lg = "english");
}
