% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_to_phonemes.R
\name{getPhon}
\alias{getPhon}
\title{Phoneme generator from distinctive features}
\usage{
getPhon(ft, lg)
}
\arguments{
\item{ft}{The abbreviated features of interest: \code{syl}, \code{son},
\code{cons}, \code{cont}, \code{DR}, \code{lat}, \code{nas}, \code{strid},
\code{vce}, \code{sg}, \code{cg}, \code{ant}, \code{cor}, \code{distr},
\code{lab}, \code{hi}, \code{lo}, \code{back}, \code{round}, \code{vel},
\code{tense}, \code{long}, \code{hitone}, \code{hireg}, \code{approx}}

\item{lg}{The language of interest: English, French, Italian, Portuguese, Spanish.
Alternatively, you can also provide your own inventory as a vector}
}
\value{
The phonemes given the features provided
}
\description{
Returns a set of phonemes given different distinctive feature values
}
\examples{
getPhon(ft = c("+hi", "+tense"), lg = "english");
getPhon(ft = c("+hi", "+tense"), lg = c("i", "u", "a", "p", "t", "k", "l", "m", "n"));
}
