% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_to_phonemes.R
\name{getPhon}
\alias{getPhon}
\title{Phoneme generator from distinctive features}
\usage{
getPhon(ft = c(), lg = "Portuguese")
}
\arguments{
\item{ft}{The abbreviated features of interest: \code{syl}, \code{son},
\code{cons}, \code{cont}, \code{DR}, \code{lat}, \code{nas}, \code{strid},
\code{vce}, \code{sg}, \code{cg}, \code{ant}, \code{cor}, \code{distr},
\code{lab}, \code{hi}, \code{lo}, \code{back}, \code{round}, \code{vel},
\code{tense}, \code{long}, \code{hitone}, \code{hireg}}

\item{lg}{The language of interest: English, French, Italian, Portuguese, Spanish}
}
\value{
The phonemes given the features provided
}
\description{
Returns a set of phonemes given different distinctive feature values
}
\examples{
getPhon(ft = c("+hi", "+tense"), lg = "english");
}
