% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_to_phonemes.R
\name{getPhon}
\alias{getPhon}
\title{Phoneme generator from distinctive features}
\usage{
getPhon(ft = c(), lg = "Portuguese")
}
\arguments{
\item{ft}{The abbreviated features of interest: syl, son, cons, cont, DR, lat, nas, strid, vce, sg, cg, ant, cor, distr, lab, hi, lo, back, round, vel, tense, long, hitone, hireg}

\item{lg}{The language of interest: English, French, Portuguese}
}
\value{
The phonemes given the features provided
}
\description{
Returns a set of phonemes given different distinctive feature values.
}
\examples{
getFeat(ft = c("+hi", "+tense"), lg = "english");
}
