% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_stress_pt.R
\name{getStress}
\alias{getStress}
\title{Stress labeller}
\usage{
getStress(word = "")
}
\arguments{
\item{word}{The string of interest using IPA phonemic transcription, already syllabified and stressed}

\item{stress}{The symbol used to mark stress}
}
\value{
The primary stress position
}
\description{
Labels a given stressed string as follows: Final, Penult, Antepenult
}
\examples{
weight_pt(word = "kom.pu.ta.ˈdoɾ", stress = "ˈ");
}
