% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSyl.R
\name{getSyl}
\alias{getSyl}
\title{Function to extract syllables from words}
\usage{
getSyl(word = "", pos = 1, syl = "\\\\.")
}
\arguments{
\item{word}{The string of interest must be syllabified}

\item{pos}{The target syllable counting from the right edge of the word}

\item{syl}{The symbol used for syllable boundaries (a period is used as the default)}
}
\value{
The desired syllable if it exists
}
\description{
Extracts a given syllable from a syllabified string
}
\examples{
getSyl(word = "kom.pu.ta.doɾ", pos = 2);
getSyl(word = "kom-pu-ta-doɾ", pos = 3, syl = "-");
}
