% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipa2tipa.R
\name{ipa2tipa}
\alias{ipa2tipa}
\title{TIPA translator}
\usage{
ipa2tipa(string)
}
\arguments{
\item{string}{The phonemically transcribed sequence from a function such as ipa()}
}
\value{
The tex code using TIPA
}
\description{
Translates a phonemically transcribed sequence into TIPA commands for LaTeX.
The function is expected to be used in conjunction with cleanText() and ipa(),
which provide the appropriate input for it
}
\examples{
ipa2tipa(string = "bo.ni.to");
}
