% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_words_pt.R
\name{ipa_pt_test}
\alias{ipa_pt_test}
\title{IPA tester for ipa_pt function in Fonology package}
\usage{
ipa_pt_test()
}
\value{
A list of words broadly and narrowly transcribed using ipa_pt
}
\description{
Returns a list of words. No arguments needed.
}
\examples{
ipa_pt_test();
}
