% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonViz.R
\name{plotSon}
\alias{plotSon}
\title{Sonority plot}
\usage{
plotSon(word = "", syl = FALSE, save_plot = FALSE)
}
\arguments{
\item{word}{A phonemically transcribed word. Syllabification and stress are optional}

\item{syl}{Whether syllables should be color-coded (requires syllabification in the input using either . or -). Defaults to F}

\item{save_plot}{Whether the plot should be saved (in your working directory). Default is F}
}
\value{
A ggplot figure with the sonority profile of the word
}
\description{
Given a string, the function plots the sonority profile of the string using
the fine-grained sonority scale in Parker (2011)
}
\examples{
plotSon(word = "tɾon.fo.ˈnil.to", syl = TRUE, save_plot = FALSE);
}
