% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vowels.R
\name{plotVowels}
\alias{plotVowels}
\title{Vowel plot generator}
\usage{
plotVowels(lg = "English", tex = F)
}
\arguments{
\item{lg}{The language whose vowel inventory will be plotted}

\item{tex}{Whether a latex file is desired. By deafult, tex = FALSE}
}
\value{
The vowel inventory desired
}
\description{
Generates vowel trapezoid using ggplot2 as well as LaTeX code using vowel package.
}
\examples{
plotVowels(lg = "portuguese");
plotVowels(lg = "english");
}
