% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_lex}
\alias{pt_lex}
\title{Portuguese Stress Lexicon - simplified}
\format{
## `pt_lex`
A data frame with 154,610 rows and 2 columns:
\describe{
  \item{word}{A word in its orthographic form}
  \item{pro}{The phonemic transcription of a word}
}
}
\source{
<https://gdgarcia.ca/psl>
}
\usage{
pt_lex
}
\description{
Comprehensive list of non-verbs in Portuguese and their phonemic transcription.
This data set is used by different functions in Fonology.
}
\keyword{datasets}
