% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_fr}
\alias{stopwords_fr}
\title{Data: stopwords in French}
\format{
A vector with stopwords in French
}
\source{
<https://cran.r-project.org/web/packages/tm/index.html>
}
\usage{
data(stopwords_fr)
}
\description{
Set of stopwords in French from the tm package.
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
