% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_pt}
\alias{stopwords_pt}
\title{Stopwords in Portuguese}
\format{
A vector with stopwords in Portuguese
}
\source{
<https://cran.r-project.org/web/packages/tm/index.html>
}
\usage{
data(stopwords_pt)
}
\description{
Set of stopwords in Portuguese from the tm package.
}
\author{
Guilherme D. Garcia (\url{https://gdgarcia.ca})
}
\keyword{datasets}
