% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords_pt}
\alias{stopwords_pt}
\title{Stopwords in Portuguese}
\format{
## `stopwords_pt`
A vector with stopwords in Portuguese
}
\source{
<https://cran.r-project.org/web/packages/tm/index.html>
}
\usage{
stopwords_pt
}
\description{
Set of stopwords in Portuguese from the tm package.
}
\keyword{datasets}
