% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe_pt.R
\name{transcribe_pt}
\alias{transcribe_pt}
\title{IPA transcriber for Portuguese}
\usage{
transcribe_pt(word = "")
}
\arguments{
\item{word}{The string of interest}
}
\value{
The IPA transcription of said string without syllabification or stress
}
\description{
Returns IPA phonemic transcription for a given string.
}
\examples{
transcribe_pt(word = "computador")
}
