% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vowels}
\alias{vowels}
\title{Hypothetical vowel formants}
\format{
## `vowels`
A hypothetical dataset with vowel formants
}
\source{
<https://gdgarcia.ca/vowels.html>
}
\usage{
vowels
}
\description{
Formant values (F1 and F2) for a 5-vowel system
}
\keyword{datasets}
