% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wug_pt.R
\name{wug_pt}
\alias{wug_pt}
\title{Word generator for Portuguese}
\usage{
wug_pt(profile = "LLL", n = 1, palatalization = F)
}
\arguments{
\item{profile}{The weight profile of the desired string using Ls or Hs}

\item{n}{Number of words to be generated}

\item{palatalization}{Whether t and d should palatalize before i (default is \code{FALSE})}
}
\value{
The IPA transcription of said string
}
\description{
Returns IPA phonemic transcription for a nonce word given a specific weight profile
}
\examples{
wug_pt(profile = "HLL", n = 2)
}
